0117_pm17:36

+++ homesweethome/src/main/java/home/staff/dao/ReceiveDAO.java
@@ -12,11 +12,11 @@ public interface ReceiveDAO {
 	
 	int selectSearchReceiveListCount(PageMaker pageMaker) throws SQLException;
 	
-	ReceiveVO selectReceiveByRcno(int rcno) throws SQLException;
+	ReceiveVO selectReceiveByRcno(long rcno) throws SQLException;
 	
 	void insertReceive(ReceiveVO receive) throws SQLException;
 	void updateReceive(ReceiveVO receive) throws SQLException;
-	void deleteReceive(int rcno) throws SQLException;
+	void deleteReceive(long rcno) throws SQLException;
 	
 	int selectReceiveSeqNext() throws SQLException;
 }
